+++ server/src/jobs-new/jobs/update-queue-metrics.job.ts
@@ -1,4 +1,4 @@
-import { Job } from '../jobs.util';
+import { Job } from '../job.class';
 
 export class UpdateQueueMetricsJob extends Job<unknown> {
   async execute() {
