--- server/src/jobs-new/startup.config.ts
@@ -1,4 +1,0 @@
-import { JobType } from './job-type.enum';
-
-// Jobs to run when the server starts
-export const STARTUP_JOBS = [JobType.SYNC_API_KEYS];
