+++ server/src/jobs-new/types/job-options.type.ts
@@ -1,0 +1,7 @@
+import { JobsOptions as BullMQJobOptions, RateLimiterOptions } from 'bullmq';
+
+export interface JobOptions extends BullMQJobOptions {
+  rateLimiter?: RateLimiterOptions;
+  skipDedupe?: boolean;
+  maxConcurrent?: number;
+}
