+++ server/src/jobs-new/types/job-type.enum.ts
